+++ webapp/frontend/app.jsx
@@ -1,32 +1,49 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
+import BottomPanel from './components/bottom-panel';
 import BucketsList from './components/buckets-list';
 import BucketContainer from './components/bucket-container';
 import Header from './components/header';
+import SettingsPage from './pages/settings';
 
 class App extends React.Component {
     constructor(props) {
         super(props);
 
         this.state = {
+            all: true,
+            apps: 0,
             currentBucket: 'Main-bucket',
-            query: ''
+            query: '',
+            page: 'buckets',
         };
         this.apiRoot = this.props.apiRoot;
+        this.changePage = this.changePage.bind(this);
         this.handleBucketChange = this.handleBucketChange.bind(this);
         this.handleSearch = this.handleSearch.bind(this);
     }
 
+    changePage(pageName) {
+        this.setState({ page: pageName });
+    }
+
     handleBucketChange(bucket) {
-        this.setState({query:'', currentBucket: bucket});
+        this.setState({ page: 'buckets', query:'', currentBucket: bucket });
     }
 
     handleSearch(newQuery) {
-        this.setState({query: newQuery});
+        this.setState({ query: newQuery });
     }
 
     render() {
+        const page = this.state.page == 'buckets' ? 
+            (<BucketContainer name={this.state.currentBucket}
+                all={this.state.all}
+                query={this.state.query}
+                handleLoad= {(apps) => this.setState({ apps })}
+                apiRoot={this.apiRoot}/>) :
+            (<SettingsPage apiRoot={this.apiRoot} />);
         return (
             <div>
                 <div className="container-fluid">
@@ -35,9 +52,10 @@ class App extends React.Component {
                             <BucketsList handleBucketChange={this.handleBucketChange} apiRoot={this.apiRoot}/>
                         </nav>
                         <main className="col-md-9 offset-md-3 pt-2">
-                            <Header onSearch={this.handleSearch}/>
-                            <BucketContainer name={this.state.currentBucket} query={this.state.query} apiRoot={this.apiRoot}/>
+                            <Header active={this.state.page} onSearch={this.handleSearch} onPageChange={this.changePage}/>
+                            { page }
                         </main>
+                        <BottomPanel apps={ this.state.apps } handleToggle= {(all) => this.setState({ all })}/>
                     </div>
                 </div>
             </div>