+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.0] - 2020-05-03
+### Added
+- Added settings page.
+- Allow showing installed packages at the top of the bucket list.
+- Added bottom panel with the number of packages and an option to show only installed.
+
 ## [v0.8] - 2019-06-13
 ### Added 
 - Make it possible to add new bucket.